+++ contracts/contracts/Storage.sol
@@ -82,7 +82,7 @@ contract Storage {
     }
 
     /// @notice Flag indicates that a user has exited certain token balance (per account id and tokenId)
-    mapping(uint24 => mapping(uint16 => bool)) public exited;
+    mapping(uint32 => mapping(uint16 => bool)) public exited;
 
     /// @notice Flag indicates that exodus (mass exit) mode is triggered
     /// @notice Once it was raised, it can not be cleared again, and all users must exit
